% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bacias_subbacias}
\alias{bacias_subbacias}
\title{Title for dataset 1}
\format{
An S4 object of class RasterLayer.
}
\usage{
bacias_subbacias
}
\description{
description for dataset1
}
\details{
details for dataset1
}
\keyword{bacia}
\keyword{subbacia}
