% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dados_estacoes_fluviometricas}
\alias{dados_estacoes_fluviometricas}
\title{Title for dataset 4}
\format{
An S4 object of class RasterLayer.
}
\usage{
dados_estacoes_fluviometricas
}
\description{
description for dataset3
}
\details{
details for dataset4
}
\keyword{dataset4}
