% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estacoes_fluviometricas}
\alias{estacoes_fluviometricas}
\title{Title for dataset 2}
\format{
An S4 object of class RasterLayer.
}
\usage{
estacoes_fluviometricas
}
\description{
description for dataset1
}
\details{
details for dataset1
}
\keyword{dataset1}
