% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pegar_estacoes_fluviometricas}
\alias{pegar_estacoes_fluviometricas}
\title{Title for dataset 3}
\format{
An S4 object of class RasterLayer.
}
\usage{
pegar_estacoes_fluviometricas
}
\description{
description for dataset3
}
\details{
details for dataset1
}
\keyword{dataset1}
